first pass at making code more testable
