+++ tests/parser.rs
@@ -1,0 +1,9 @@
+use graphler::{
+    parser::{parse}
+};
+
+#[test]
+fn valid_collection_json() {
+    // todo: break up parser into parseing and file creation
+    // test parser output against a mocked valid postman.collection.json
+}
